Fix docs.

+++ cargo_util.py
@@ -5,7 +5,7 @@ import fileinput
 import re
 import os
 import glob
-from shutil import copy
+from shutil import copy, rmtree
 from subprocess import check_call
 
 crate_list="""
@@ -85,12 +85,18 @@ if args.clean:
 		check_call(cargo_cmd('clean'), cwd=crate)
 
 if args.doc:
+	rmtree('doc/target/doc')
 	print 'Building docs'
-	check_call(cargo_cmd('doc'), cwd='doc')
+	check_call(['cargo', 'doc'], cwd='doc')
 	print 'Fixing up the search index'
+	found = False
 	for line in fileinput.input('doc/target/doc/search-index.js', inplace=1):
-		new_line = re.sub(r"searchIndex\['delete_me'\].*", '', line)
+		new_line = re.sub(r'searchIndex\["delete_me"\].*', '', line)
+		if new_line != line:
+			found = True
 		print new_line,
+	if not found:
+		raise Exception("Couldn't find the line in search-index.js!")
 	print 'Copying new CSS'
 	for path in glob.glob('doc/target/doc/*.css'):
 		os.remove(path)
